requirements.txt
asgiref==3.11.1
brotli==1.2.0
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djlint==1.36.4
EditorConfig==0.17.1
gunicorn==25.0.
h11==0.16.0
jsbeautifier==1.15.4
json5==0.13.0
packaging==26.0
pathspec==1.0.4
PyJWT==2.11.0
PyYAML==6.0.3
regex==2026.1.15
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
uvicorn==0.40.0
whitenoise==6.11.0
